src/api/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
random


pyspark
nltk
numpy
